[DESIGN/#26] Refactor: 게시물 및 답변 관련 API 경로 정리 및 파일 이름 수정
