+++ app/api/boards/route.ts
@@ -1,5 +1,5 @@
 import { NextRequest, NextResponse } from "next/server";
-import GetGeneralPosts from "@/utils/post/general";
+import GetPostList from "@/utils/post/list";
 
 export async function GET(req: NextRequest) {
   // request-header로 access_token 받기
@@ -8,6 +8,6 @@ export async function GET(req: NextRequest) {
     return NextResponse.json({ error: "no_access_token" }, { status: 401 });
   }
 
-  const posts = await GetGeneralPosts(accessToken);
+  const posts = await GetPostList(accessToken);
   return NextResponse.json(posts);
 }
